Development

+++ src/server/api/routers/sectors.ts
@@ -5,6 +5,7 @@ import { Ratelimit } from "@upstash/ratelimit";
 import { Redis } from "@upstash/redis";
 import { TRPCError } from "@trpc/server";
 import { type Prisma } from "@prisma/client";
+import { clerkClient } from "@clerk/nextjs";
 
 const redis = new Redis({
   url: "https://us1-merry-snake-32728.upstash.io",
@@ -23,6 +24,14 @@ export const sectorsRouter = createTRPCRouter({
       const sector = await ctx.prisma.sector.findUnique({
         where: {
           id: input.id,
+        },
+        include: {
+          _count: {
+            select: {
+              CrewMembers: true,
+              Projects: true,
+            },
+          },
         },
       });
 
@@ -75,9 +84,15 @@ export const sectorsRouter = createTRPCRouter({
   create: privateProcedure
     .input(
       z.object({
-        name: z.string(),
+        name: z
+          .string()
+          .min(3, "Name must be at least 3 characters long")
+          .max(50, "Name must be at most 50 characters long"),
         description: z.string(),
-        departmentCode: z.string(),
+        departmentCode: z
+          .string()
+          .min(2, "Code must be at least 2 characters long")
+          .max(10, "Code must be at most 10 characters long"),
         color: z.string(),
       })
     )
@@ -91,16 +106,27 @@ export const sectorsRouter = createTRPCRouter({
         });
       }
 
-      const sectorExists = await ctx.prisma.sector.findUnique({
+      const sectorExists = await ctx.prisma.sector.findFirst({
         where: {
-          name: input.name,
+          OR: [
+            {
+              name: {
+                equals: input.name,
+              },
+            },
+            {
+              departmentCode: {
+                equals: input.departmentCode,
+              },
+            },
+          ],
         },
       });
 
       if (sectorExists) {
         throw new TRPCError({
           code: "BAD_REQUEST",
-          message: "A sector with this name already exists",
+          message: "A sector with the name and/or code already exists",
         });
       }
 
@@ -129,9 +155,16 @@ export const sectorsRouter = createTRPCRouter({
     .input(
       z.object({
         id: z.string(),
-        name: z.string(),
+        name: z
+          .string()
+          .min(3, "Name must be at least 3 characters long")
+          .max(50, "Name must be at most 50 characters long"),
         description: z.string(),
-        departmentCode: z.string(),
+        departmentCode: z
+          .string()
+          .min(2, "Code must be at least 2 characters long")
+          .max(10, "Code must be at most 10 characters long"),
+        color: z.string(),
       })
     )
     .mutation(async ({ ctx, input }) => {
@@ -149,6 +182,14 @@ export const sectorsRouter = createTRPCRouter({
       const sector = await ctx.prisma.sector.update({
         where: {
           id: input.id,
+        },
+        include: {
+          _count: {
+            select: {
+              CrewMembers: true,
+              Projects: true,
+            },
+          },
         },
         data: {
           name: input.name,
@@ -175,9 +216,56 @@ export const sectorsRouter = createTRPCRouter({
         });
       }
 
+      const user = await clerkClient.users.getUser(authorId);
+
+      const email = user?.emailAddresses[0]?.emailAddress;
+
+      if (!user || !email || user.banned) {
+        throw new TRPCError({
+          code: "UNAUTHORIZED",
+          message: "You are not authorized to perform this action",
+        });
+      }
+
+      const getSector = await ctx.prisma.sector.findUnique({
+        where: {
+          id: input.id,
+        },
+        include: {
+          _count: {
+            select: {
+              CrewMembers: true,
+              Projects: true,
+            },
+          },
+        },
+      });
+
+      if (
+        getSector?._count.CrewMembers !== 0 ||
+        getSector?._count.Projects !== 0
+      ) {
+        throw new TRPCError({
+          code: "BAD_REQUEST",
+          message: "You cannot delete a sector with crew members or projects",
+        });
+      }
+
       const sector = await ctx.prisma.sector.delete({
         where: {
           id: input.id,
+        },
+      });
+
+      await ctx.prisma.log.create({
+        data: {
+          action: "none",
+          category: "crew",
+          name: `Deleted Sector ${sector.name}`,
+          authorId: authorId,
+          url: `/#`,
+          description: `${email} deleted sector ${sector.name} (department ${sector.departmentCode})`,
+          severity: "critical",
         },
       });
